refactor: move buildStepValidator to new path and add option to validate multiple fields
