+++ app/routes/shared/formular.server.ts
@@ -239,7 +239,7 @@ export const action = async ({ request }: ActionFunctionArgs) => {
     }
   }
 
-  const validationResult = await validateFormData(flowId, relevantFormData);
+  const validationResult = await validateFormData(pathname, relevantFormData);
 
   if (validationResult.error)
     return validationError(
